pyproject.toml
[project]
name = "oap-nextgen"
version = "0.2.0"
description = "A MILP and Benders-Decomposition framework for the Optimal Area Polygonization problem."
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "cplex>=22.1.2.0",
    "docplex>=2.31.254",
    "gurobipy>=13.0.0",
    "inquirer>=3.4.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pandas-stubs>=3.0.0.260204",
    "plotly>=6.5.2",
    "pycddlib<3.0.0",
    "pytest>=9.0.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "umap-learn>=0.5.11",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
# E501  – line-too-long: solver constraint expressions routinely exceed any sane limit
# N999  – PascalCase module names: renaming OAPBaseModel.py etc. breaks 20+ import sites
# N802/N803/N806 – math-convention names (A_pp, D, G, TOL, CH, S_left …) are domain standard
# SIM   – all SIM style rules: research codebase, cosmetic churn not worth it
# UP015 – unnecessary open() mode 'r': cosmetic
ignore = ["E501", "N999", "N802", "N803", "N806", "SIM", "UP015"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
